Add escalation routing tests for memory-graph suggest.
Cover Quick success, Brain escalation paths, dual failure, RDF non-escalation, and approve-route LLM isolation.
